+++ src/Http/Session.php
@@ -17,7 +17,7 @@ namespace Origin\Http;
 
 use Origin\Core\Dot;
 use Origin\Core\Config;
-use Origin\Utility\Security;
+use Origin\Security\Security;
 use Origin\Exception\Exception;
 
 class Session