+++ tests/TestCase/Http/SessionTest.php
@@ -15,7 +15,7 @@
 namespace Origin\Test\Http;
 
 use Origin\Http\Session;
-use Origin\Utility\Security;
+use Origin\Security\Security;
 use Origin\TestSuite\TestTrait;
 
 class MockSession extends Session