+++ tests/TestCase/Model/ModelTest.php
@@ -23,7 +23,7 @@ use ArrayObject;
 use Origin\Model\Entity;
 
 use Origin\Model\Collection;
-use Origin\Utility\Security;
+use Origin\Security\Security;
 use Origin\Exception\Exception;
 use Origin\Model\ModelRegistry;
 use Origin\Model\Model as BaseModel;